wp12_hackathon/dissemination_summary_prototype/requirements.txt
langchain>=0.3.10
langchain-community>=0.0.10
langchain-core>=0.1.10
python-dotenv>=1.0.0
langchain-openai>=0.0.2
docling>=0.2.0
openai>=1.0.0
langchain-ollama>=0.0.3 
pypdf>=5.6
logger>=1.4
langchain-docling
sentence-transformers
langchain-chroma
langchain-huggingface
chromadb
nicegui
starlette